move template strings to submodule
